% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_api.R
\docType{class}
\name{DefaultApi}
\alias{DefaultApi}
\title{Default operations}
\format{
An \code{R6Class} generator object
}
\description{
DefaultApi
}
\details{
FastAPI

No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

The version of the OpenAPI document: 0.1.0
Generated by: https://openapi-generator.tech
}
\examples{
\dontrun{
####################  AddCollaboratorPut  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_account_id <- "account_id_example" # character |
var_schema_id <- "schema_id_example" # character |
var_role <- "read" # character |  (Optional)
var_authorization <- "authorization_example" # character |  (Optional)

# Add Collaborator
api_instance <- DefaultApi$new()



####################  CreateInstanceInstancesPut  ####################

library(laminr.api)
var_name <- "name_example" # character |
var_storage <- "create-s3" # character |  (Optional)
var_schema_str <- "schema_str_example" # character |  (Optional)
var_db_server_key <- "db_server_key_example" # character |  (Optional)
var_db_server_url <- "db_server_url_example" # character |  (Optional)
var_authorization <- "authorization_example" # character |  (Optional)

# Create Instance
api_instance <- DefaultApi$new()



####################  CreateRecordPut  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_module_name <- "module_name_example" # character |
var_model_name <- "model_name_example" # character |
var_schema_id <- "schema_id_example" # character |
var_body <- TODO # AnyType |
var_authorization <- "authorization_example" # character |  (Optional)

# Create Record
api_instance <- DefaultApi$new()



####################  DeleteCollaboratorDelete  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_account_id <- "account_id_example" # character |
var_authorization <- "authorization_example" # character |  (Optional)

# Delete Collaborator
api_instance <- DefaultApi$new()



####################  DeleteInstanceDelete  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_instance_name <- "instance_name_example" # character |
var_authorization <- "authorization_example" # character |  (Optional)

# Delete Instance
api_instance <- DefaultApi$new()



####################  DeleteRecordDelete  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_module_name <- "module_name_example" # character |
var_model_name <- "model_name_example" # character |
var_uid <- "uid_example" # character |
var_schema_id <- "schema_id_example" # character |
var_authorization <- "authorization_example" # character |  (Optional)

# Delete Record
api_instance <- DefaultApi$new()



####################  GeneratePasswordGet  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_authorization <- "authorization_example" # character |  (Optional)

# Generate Password
api_instance <- DefaultApi$new()



####################  GenerateUrlGet  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_authorization <- "authorization_example" # character |  (Optional)

# Generate Url
api_instance <- DefaultApi$new()



####################  GetInstanceStatisticsGet  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_schema_id <- "schema_id_example" # character |
var_q <- c("inner_example") # array[character] | In module + "." + model format (case-sensitive) (Optional)
var_authorization <- "authorization_example" # character |  (Optional)

# Get Instance Statistics
api_instance <- DefaultApi$new()



####################  GetNonEmptyTablesGet  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_schema_id <- "schema_id_example" # character |
var_authorization <- "authorization_example" # character |  (Optional)

# Get Non Empty Tables
api_instance <- DefaultApi$new()



####################  GetRecordPost  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_module_name <- "module_name_example" # character |
var_model_name <- "model_name_example" # character |
var_id_or_uid <- "id_or_uid_example" # character |
var_schema_id <- "schema_id_example" # character |
var_limit_to_many <- 10 # integer |  (Optional)
var_include_foreign_keys <- FALSE # character |  (Optional)
var_authorization <- "authorization_example" # character |  (Optional)
var_get_record_request_body <- GetRecordRequestBody$new(c("select_example")) # GetRecordRequestBody |  (Optional)

# Get Record
api_instance <- DefaultApi$new()



####################  GetRecordsPost  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_module_name <- "module_name_example" # character |
var_model_name <- "model_name_example" # character |
var_schema_id <- "schema_id_example" # character |
var_limit <- 50 # integer |  (Optional)
var_offset <- 0 # integer |  (Optional)
var_limit_to_many <- 10 # integer |  (Optional)
var_include_foreign_keys <- FALSE # character |  (Optional)
var_authorization <- "authorization_example" # character |  (Optional)
var_get_records_request_body <- GetRecordsRequestBody$new(c("select_example"), 123, c(OrderByColumn$new("field_example", "descending_example")), "search_example") # GetRecordsRequestBody |  (Optional)

# Get Records
api_instance <- DefaultApi$new()



####################  GetRelationCountsGet  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_module_name <- "module_name_example" # character |
var_model_name <- "model_name_example" # character |
var_id <- 56 # integer |
var_schema_id <- "schema_id_example" # character |
var_authorization <- "authorization_example" # character |  (Optional)

# Get Relation Counts
api_instance <- DefaultApi$new()



####################  GetRelationsGet  ####################

library(laminr.api)
var_module_name <- "module_name_example" # character |
var_model_name <- "model_name_example" # character |
var_instance_id <- "instance_id_example" # character |
var_authorization <- "authorization_example" # character |  (Optional)

# Get Relations
api_instance <- DefaultApi$new()



####################  GetSchemaGet  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_authorization <- "authorization_example" # character |  (Optional)

# Get Schema
api_instance <- DefaultApi$new()



####################  GetValuesPost  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_module_name <- "module_name_example" # character |
var_model_name <- "model_name_example" # character |
var_field_path <- "field_path_example" # character |
var_schema_id <- "schema_id_example" # character |
var_limit <- 50 # integer |  (Optional)
var_offset <- 0 # integer |  (Optional)
var_authorization <- "authorization_example" # character |  (Optional)
var_get_values_request_body <- GetValuesRequestBody$new(123, c(OrderByColumn$new("field_example", "descending_example")), "search_example") # GetValuesRequestBody |  (Optional)

# Get Values
api_instance <- DefaultApi$new()



####################  GrantS3PermissionsPut  ####################

library(laminr.api)
var_bucket_name <- "bucket_name_example" # character |
var_s3_permissions_request <- S3PermissionsRequest$new("aws_access_key_id_example", "aws_secret_access_key_example", "region_example") # S3PermissionsRequest |
var_aws_account_id <- "767398070972" # character |  (Optional)
var_aws_user_name <- "lamin-manager" # character |  (Optional)
var_authorization <- "authorization_example" # character |  (Optional)

# Grant S3 Permissions
api_instance <- DefaultApi$new()



####################  GroupByPost  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_module_name <- "module_name_example" # character |
var_model_name <- "model_name_example" # character |
var_schema_id <- "schema_id_example" # character |
var_group_by_request_body <- GroupByRequestBody$new(c(Dimension$new("field_name_example", "count")), c(Measure$new("field_name_example", "count", "alias_example")), 123) # GroupByRequestBody |
var_authorization <- "authorization_example" # character |  (Optional)

# Group By
api_instance <- DefaultApi$new()



####################  TransferOwnershipPatch  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_handle <- "handle_example" # character |
var_schema_id <- "schema_id_example" # character |
var_authorization <- "authorization_example" # character |  (Optional)

# Transfer Ownership
api_instance <- DefaultApi$new()



####################  UpdateCollaboratorPatch  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_account_id <- "account_id_example" # character |
var_role <- "role_example" # character |
var_authorization <- "authorization_example" # character |  (Optional)

# Update Collaborator
api_instance <- DefaultApi$new()



####################  UpdateRecordPatch  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_module_name <- "module_name_example" # character |
var_model_name <- "model_name_example" # character |
var_uid <- "uid_example" # character |
var_schema_id <- "schema_id_example" # character |
var_body <- TODO # AnyType |
var_authorization <- "authorization_example" # character |  (Optional)

# Update Record
api_instance <- DefaultApi$new()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DefaultApi-new}{\code{DefaultApi$new()}}
\item \href{#method-DefaultApi-AddCollaboratorPut}{\code{DefaultApi$AddCollaboratorPut()}}
\item \href{#method-DefaultApi-AddCollaboratorPutWithHttpInfo}{\code{DefaultApi$AddCollaboratorPutWithHttpInfo()}}
\item \href{#method-DefaultApi-CreateInstanceInstancesPut}{\code{DefaultApi$CreateInstanceInstancesPut()}}
\item \href{#method-DefaultApi-CreateInstanceInstancesPutWithHttpInfo}{\code{DefaultApi$CreateInstanceInstancesPutWithHttpInfo()}}
\item \href{#method-DefaultApi-CreateRecordPut}{\code{DefaultApi$CreateRecordPut()}}
\item \href{#method-DefaultApi-CreateRecordPutWithHttpInfo}{\code{DefaultApi$CreateRecordPutWithHttpInfo()}}
\item \href{#method-DefaultApi-DeleteCollaboratorDelete}{\code{DefaultApi$DeleteCollaboratorDelete()}}
\item \href{#method-DefaultApi-DeleteCollaboratorDeleteWithHttpInfo}{\code{DefaultApi$DeleteCollaboratorDeleteWithHttpInfo()}}
\item \href{#method-DefaultApi-DeleteInstanceDelete}{\code{DefaultApi$DeleteInstanceDelete()}}
\item \href{#method-DefaultApi-DeleteInstanceDeleteWithHttpInfo}{\code{DefaultApi$DeleteInstanceDeleteWithHttpInfo()}}
\item \href{#method-DefaultApi-DeleteRecordDelete}{\code{DefaultApi$DeleteRecordDelete()}}
\item \href{#method-DefaultApi-DeleteRecordDeleteWithHttpInfo}{\code{DefaultApi$DeleteRecordDeleteWithHttpInfo()}}
\item \href{#method-DefaultApi-GeneratePasswordGet}{\code{DefaultApi$GeneratePasswordGet()}}
\item \href{#method-DefaultApi-GeneratePasswordGetWithHttpInfo}{\code{DefaultApi$GeneratePasswordGetWithHttpInfo()}}
\item \href{#method-DefaultApi-GenerateUrlGet}{\code{DefaultApi$GenerateUrlGet()}}
\item \href{#method-DefaultApi-GenerateUrlGetWithHttpInfo}{\code{DefaultApi$GenerateUrlGetWithHttpInfo()}}
\item \href{#method-DefaultApi-GetInstanceStatisticsGet}{\code{DefaultApi$GetInstanceStatisticsGet()}}
\item \href{#method-DefaultApi-GetInstanceStatisticsGetWithHttpInfo}{\code{DefaultApi$GetInstanceStatisticsGetWithHttpInfo()}}
\item \href{#method-DefaultApi-GetNonEmptyTablesGet}{\code{DefaultApi$GetNonEmptyTablesGet()}}
\item \href{#method-DefaultApi-GetNonEmptyTablesGetWithHttpInfo}{\code{DefaultApi$GetNonEmptyTablesGetWithHttpInfo()}}
\item \href{#method-DefaultApi-GetRecordPost}{\code{DefaultApi$GetRecordPost()}}
\item \href{#method-DefaultApi-GetRecordPostWithHttpInfo}{\code{DefaultApi$GetRecordPostWithHttpInfo()}}
\item \href{#method-DefaultApi-GetRecordsPost}{\code{DefaultApi$GetRecordsPost()}}
\item \href{#method-DefaultApi-GetRecordsPostWithHttpInfo}{\code{DefaultApi$GetRecordsPostWithHttpInfo()}}
\item \href{#method-DefaultApi-GetRelationCountsGet}{\code{DefaultApi$GetRelationCountsGet()}}
\item \href{#method-DefaultApi-GetRelationCountsGetWithHttpInfo}{\code{DefaultApi$GetRelationCountsGetWithHttpInfo()}}
\item \href{#method-DefaultApi-GetRelationsGet}{\code{DefaultApi$GetRelationsGet()}}
\item \href{#method-DefaultApi-GetRelationsGetWithHttpInfo}{\code{DefaultApi$GetRelationsGetWithHttpInfo()}}
\item \href{#method-DefaultApi-GetSchemaGet}{\code{DefaultApi$GetSchemaGet()}}
\item \href{#method-DefaultApi-GetSchemaGetWithHttpInfo}{\code{DefaultApi$GetSchemaGetWithHttpInfo()}}
\item \href{#method-DefaultApi-GetValuesPost}{\code{DefaultApi$GetValuesPost()}}
\item \href{#method-DefaultApi-GetValuesPostWithHttpInfo}{\code{DefaultApi$GetValuesPostWithHttpInfo()}}
\item \href{#method-DefaultApi-GrantS3PermissionsPut}{\code{DefaultApi$GrantS3PermissionsPut()}}
\item \href{#method-DefaultApi-GrantS3PermissionsPutWithHttpInfo}{\code{DefaultApi$GrantS3PermissionsPutWithHttpInfo()}}
\item \href{#method-DefaultApi-GroupByPost}{\code{DefaultApi$GroupByPost()}}
\item \href{#method-DefaultApi-GroupByPostWithHttpInfo}{\code{DefaultApi$GroupByPostWithHttpInfo()}}
\item \href{#method-DefaultApi-TransferOwnershipPatch}{\code{DefaultApi$TransferOwnershipPatch()}}
\item \href{#method-DefaultApi-TransferOwnershipPatchWithHttpInfo}{\code{DefaultApi$TransferOwnershipPatchWithHttpInfo()}}
\item \href{#method-DefaultApi-UpdateCollaboratorPatch}{\code{DefaultApi$UpdateCollaboratorPatch()}}
\item \href{#method-DefaultApi-UpdateCollaboratorPatchWithHttpInfo}{\code{DefaultApi$UpdateCollaboratorPatchWithHttpInfo()}}
\item \href{#method-DefaultApi-UpdateRecordPatch}{\code{DefaultApi$UpdateRecordPatch()}}
\item \href{#method-DefaultApi-UpdateRecordPatchWithHttpInfo}{\code{DefaultApi$UpdateRecordPatchWithHttpInfo()}}
\item \href{#method-DefaultApi-clone}{\code{DefaultApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DefaultApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-AddCollaboratorPut"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-AddCollaboratorPut}{}}}
\subsection{Method \code{AddCollaboratorPut()}}{
Add Collaborator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$AddCollaboratorPut(
  instance_id,
  account_id,
  schema_id,
  role = "read",
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{role}}{(optional) No description (default value: "read")}

\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-AddCollaboratorPutWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-AddCollaboratorPutWithHttpInfo}{}}}
\subsection{Method \code{AddCollaboratorPutWithHttpInfo()}}{
Add Collaborator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$AddCollaboratorPutWithHttpInfo(
  instance_id,
  account_id,
  schema_id,
  role = "read",
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{role}}{(optional) No description (default value: "read")}

\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-CreateInstanceInstancesPut"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-CreateInstanceInstancesPut}{}}}
\subsection{Method \code{CreateInstanceInstancesPut()}}{
Create Instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$CreateInstanceInstancesPut(
  name,
  storage = "create-s3",
  schema_str = NULL,
  db_server_key = NULL,
  db_server_url = NULL,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{storage}}{(optional) No description (default value: "create-s3")}

\item{\code{schema_str}}{(optional) No description}

\item{\code{db_server_key}}{(optional) No description}

\item{\code{db_server_url}}{(optional) No description}

\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-CreateInstanceInstancesPutWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-CreateInstanceInstancesPutWithHttpInfo}{}}}
\subsection{Method \code{CreateInstanceInstancesPutWithHttpInfo()}}{
Create Instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$CreateInstanceInstancesPutWithHttpInfo(
  name,
  storage = "create-s3",
  schema_str = NULL,
  db_server_key = NULL,
  db_server_url = NULL,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{storage}}{(optional) No description (default value: "create-s3")}

\item{\code{schema_str}}{(optional) No description}

\item{\code{db_server_key}}{(optional) No description}

\item{\code{db_server_url}}{(optional) No description}

\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-CreateRecordPut"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-CreateRecordPut}{}}}
\subsection{Method \code{CreateRecordPut()}}{
Create Record
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$CreateRecordPut(
  instance_id,
  module_name,
  model_name,
  schema_id,
  body,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-CreateRecordPutWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-CreateRecordPutWithHttpInfo}{}}}
\subsection{Method \code{CreateRecordPutWithHttpInfo()}}{
Create Record
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$CreateRecordPutWithHttpInfo(
  instance_id,
  module_name,
  model_name,
  schema_id,
  body,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-DeleteCollaboratorDelete"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-DeleteCollaboratorDelete}{}}}
\subsection{Method \code{DeleteCollaboratorDelete()}}{
Delete Collaborator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$DeleteCollaboratorDelete(
  instance_id,
  account_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-DeleteCollaboratorDeleteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-DeleteCollaboratorDeleteWithHttpInfo}{}}}
\subsection{Method \code{DeleteCollaboratorDeleteWithHttpInfo()}}{
Delete Collaborator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$DeleteCollaboratorDeleteWithHttpInfo(
  instance_id,
  account_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-DeleteInstanceDelete"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-DeleteInstanceDelete}{}}}
\subsection{Method \code{DeleteInstanceDelete()}}{
Delete Instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$DeleteInstanceDelete(
  instance_id,
  instance_name,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-DeleteInstanceDeleteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-DeleteInstanceDeleteWithHttpInfo}{}}}
\subsection{Method \code{DeleteInstanceDeleteWithHttpInfo()}}{
Delete Instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$DeleteInstanceDeleteWithHttpInfo(
  instance_id,
  instance_name,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-DeleteRecordDelete"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-DeleteRecordDelete}{}}}
\subsection{Method \code{DeleteRecordDelete()}}{
Delete Record
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$DeleteRecordDelete(
  instance_id,
  module_name,
  model_name,
  uid,
  schema_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-DeleteRecordDeleteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-DeleteRecordDeleteWithHttpInfo}{}}}
\subsection{Method \code{DeleteRecordDeleteWithHttpInfo()}}{
Delete Record
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$DeleteRecordDeleteWithHttpInfo(
  instance_id,
  module_name,
  model_name,
  uid,
  schema_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GeneratePasswordGet"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GeneratePasswordGet}{}}}
\subsection{Method \code{GeneratePasswordGet()}}{
Generate Password
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GeneratePasswordGet(
  instance_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GeneratePasswordGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GeneratePasswordGetWithHttpInfo}{}}}
\subsection{Method \code{GeneratePasswordGetWithHttpInfo()}}{
Generate Password
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GeneratePasswordGetWithHttpInfo(
  instance_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GenerateUrlGet"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GenerateUrlGet}{}}}
\subsection{Method \code{GenerateUrlGet()}}{
Generate Url
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GenerateUrlGet(
  instance_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GenerateUrlGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GenerateUrlGetWithHttpInfo}{}}}
\subsection{Method \code{GenerateUrlGetWithHttpInfo()}}{
Generate Url
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GenerateUrlGetWithHttpInfo(
  instance_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetInstanceStatisticsGet"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetInstanceStatisticsGet}{}}}
\subsection{Method \code{GetInstanceStatisticsGet()}}{
Get Instance Statistics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetInstanceStatisticsGet(
  instance_id,
  schema_id,
  q = NULL,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q}}{(optional) In module + "." + model format (case-sensitive)}

\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetInstanceStatisticsGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetInstanceStatisticsGetWithHttpInfo}{}}}
\subsection{Method \code{GetInstanceStatisticsGetWithHttpInfo()}}{
Get Instance Statistics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetInstanceStatisticsGetWithHttpInfo(
  instance_id,
  schema_id,
  q = NULL,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q}}{(optional) In module + "." + model format (case-sensitive)}

\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetNonEmptyTablesGet"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetNonEmptyTablesGet}{}}}
\subsection{Method \code{GetNonEmptyTablesGet()}}{
Get Non Empty Tables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetNonEmptyTablesGet(
  instance_id,
  schema_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetNonEmptyTablesGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetNonEmptyTablesGetWithHttpInfo}{}}}
\subsection{Method \code{GetNonEmptyTablesGetWithHttpInfo()}}{
Get Non Empty Tables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetNonEmptyTablesGetWithHttpInfo(
  instance_id,
  schema_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetRecordPost"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetRecordPost}{}}}
\subsection{Method \code{GetRecordPost()}}{
Get Record
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetRecordPost(
  instance_id,
  module_name,
  model_name,
  id_or_uid,
  schema_id,
  limit_to_many = 10,
  include_foreign_keys = FALSE,
  authorization = NULL,
  get_record_request_body = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit_to_many}}{(optional) No description (default value: 10)}

\item{\code{include_foreign_keys}}{(optional) No description (default value: FALSE)}

\item{\code{authorization}}{(optional) No description}

\item{\code{get_record_request_body}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetRecordPostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetRecordPostWithHttpInfo}{}}}
\subsection{Method \code{GetRecordPostWithHttpInfo()}}{
Get Record
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetRecordPostWithHttpInfo(
  instance_id,
  module_name,
  model_name,
  id_or_uid,
  schema_id,
  limit_to_many = 10,
  include_foreign_keys = FALSE,
  authorization = NULL,
  get_record_request_body = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit_to_many}}{(optional) No description (default value: 10)}

\item{\code{include_foreign_keys}}{(optional) No description (default value: FALSE)}

\item{\code{authorization}}{(optional) No description}

\item{\code{get_record_request_body}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetRecordsPost"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetRecordsPost}{}}}
\subsection{Method \code{GetRecordsPost()}}{
Get Records
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetRecordsPost(
  instance_id,
  module_name,
  model_name,
  schema_id,
  limit = 50,
  offset = 0,
  limit_to_many = 10,
  include_foreign_keys = FALSE,
  authorization = NULL,
  get_records_request_body = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{(optional) No description (default value: 50)}

\item{\code{offset}}{(optional) No description (default value: 0)}

\item{\code{limit_to_many}}{(optional) No description (default value: 10)}

\item{\code{include_foreign_keys}}{(optional) No description (default value: FALSE)}

\item{\code{authorization}}{(optional) No description}

\item{\code{get_records_request_body}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetRecordsPostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetRecordsPostWithHttpInfo}{}}}
\subsection{Method \code{GetRecordsPostWithHttpInfo()}}{
Get Records
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetRecordsPostWithHttpInfo(
  instance_id,
  module_name,
  model_name,
  schema_id,
  limit = 50,
  offset = 0,
  limit_to_many = 10,
  include_foreign_keys = FALSE,
  authorization = NULL,
  get_records_request_body = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{(optional) No description (default value: 50)}

\item{\code{offset}}{(optional) No description (default value: 0)}

\item{\code{limit_to_many}}{(optional) No description (default value: 10)}

\item{\code{include_foreign_keys}}{(optional) No description (default value: FALSE)}

\item{\code{authorization}}{(optional) No description}

\item{\code{get_records_request_body}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetRelationCountsGet"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetRelationCountsGet}{}}}
\subsection{Method \code{GetRelationCountsGet()}}{
Get Relation Counts
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetRelationCountsGet(
  instance_id,
  module_name,
  model_name,
  id,
  schema_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetRelationCountsGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetRelationCountsGetWithHttpInfo}{}}}
\subsection{Method \code{GetRelationCountsGetWithHttpInfo()}}{
Get Relation Counts
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetRelationCountsGetWithHttpInfo(
  instance_id,
  module_name,
  model_name,
  id,
  schema_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetRelationsGet"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetRelationsGet}{}}}
\subsection{Method \code{GetRelationsGet()}}{
Get Relations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetRelationsGet(
  module_name,
  model_name,
  instance_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetRelationsGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetRelationsGetWithHttpInfo}{}}}
\subsection{Method \code{GetRelationsGetWithHttpInfo()}}{
Get Relations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetRelationsGetWithHttpInfo(
  module_name,
  model_name,
  instance_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetSchemaGet"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetSchemaGet}{}}}
\subsection{Method \code{GetSchemaGet()}}{
Get Schema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetSchemaGet(
  instance_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetSchemaGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetSchemaGetWithHttpInfo}{}}}
\subsection{Method \code{GetSchemaGetWithHttpInfo()}}{
Get Schema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetSchemaGetWithHttpInfo(
  instance_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetValuesPost"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetValuesPost}{}}}
\subsection{Method \code{GetValuesPost()}}{
Get Values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetValuesPost(
  instance_id,
  module_name,
  model_name,
  field_path,
  schema_id,
  limit = 50,
  offset = 0,
  authorization = NULL,
  get_values_request_body = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{(optional) No description (default value: 50)}

\item{\code{offset}}{(optional) No description (default value: 0)}

\item{\code{authorization}}{(optional) No description}

\item{\code{get_values_request_body}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetValuesPostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetValuesPostWithHttpInfo}{}}}
\subsection{Method \code{GetValuesPostWithHttpInfo()}}{
Get Values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetValuesPostWithHttpInfo(
  instance_id,
  module_name,
  model_name,
  field_path,
  schema_id,
  limit = 50,
  offset = 0,
  authorization = NULL,
  get_values_request_body = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{(optional) No description (default value: 50)}

\item{\code{offset}}{(optional) No description (default value: 0)}

\item{\code{authorization}}{(optional) No description}

\item{\code{get_values_request_body}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GrantS3PermissionsPut"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GrantS3PermissionsPut}{}}}
\subsection{Method \code{GrantS3PermissionsPut()}}{
Grant S3 Permissions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GrantS3PermissionsPut(
  bucket_name,
  s3_permissions_request,
  aws_account_id = "767398070972",
  aws_user_name = "lamin-manager",
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aws_account_id}}{(optional) No description (default value: "767398070972")}

\item{\code{aws_user_name}}{(optional) No description (default value: "lamin-manager")}

\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GrantS3PermissionsPutWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GrantS3PermissionsPutWithHttpInfo}{}}}
\subsection{Method \code{GrantS3PermissionsPutWithHttpInfo()}}{
Grant S3 Permissions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GrantS3PermissionsPutWithHttpInfo(
  bucket_name,
  s3_permissions_request,
  aws_account_id = "767398070972",
  aws_user_name = "lamin-manager",
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aws_account_id}}{(optional) No description (default value: "767398070972")}

\item{\code{aws_user_name}}{(optional) No description (default value: "lamin-manager")}

\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GroupByPost"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GroupByPost}{}}}
\subsection{Method \code{GroupByPost()}}{
Group By
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GroupByPost(
  instance_id,
  module_name,
  model_name,
  schema_id,
  group_by_request_body,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GroupByPostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GroupByPostWithHttpInfo}{}}}
\subsection{Method \code{GroupByPostWithHttpInfo()}}{
Group By
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GroupByPostWithHttpInfo(
  instance_id,
  module_name,
  model_name,
  schema_id,
  group_by_request_body,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-TransferOwnershipPatch"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-TransferOwnershipPatch}{}}}
\subsection{Method \code{TransferOwnershipPatch()}}{
Transfer Ownership
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$TransferOwnershipPatch(
  instance_id,
  handle,
  schema_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-TransferOwnershipPatchWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-TransferOwnershipPatchWithHttpInfo}{}}}
\subsection{Method \code{TransferOwnershipPatchWithHttpInfo()}}{
Transfer Ownership
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$TransferOwnershipPatchWithHttpInfo(
  instance_id,
  handle,
  schema_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-UpdateCollaboratorPatch"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-UpdateCollaboratorPatch}{}}}
\subsection{Method \code{UpdateCollaboratorPatch()}}{
Update Collaborator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$UpdateCollaboratorPatch(
  instance_id,
  account_id,
  role,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-UpdateCollaboratorPatchWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-UpdateCollaboratorPatchWithHttpInfo}{}}}
\subsection{Method \code{UpdateCollaboratorPatchWithHttpInfo()}}{
Update Collaborator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$UpdateCollaboratorPatchWithHttpInfo(
  instance_id,
  account_id,
  role,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-UpdateRecordPatch"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-UpdateRecordPatch}{}}}
\subsection{Method \code{UpdateRecordPatch()}}{
Update Record
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$UpdateRecordPatch(
  instance_id,
  module_name,
  model_name,
  uid,
  schema_id,
  body,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-UpdateRecordPatchWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-UpdateRecordPatchWithHttpInfo}{}}}
\subsection{Method \code{UpdateRecordPatchWithHttpInfo()}}{
Update Record
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$UpdateRecordPatchWithHttpInfo(
  instance_id,
  module_name,
  model_name,
  uid,
  schema_id,
  body,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
