% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_record_request_body.R
\docType{class}
\name{GetRecordRequestBody}
\alias{GetRecordRequestBody}
\title{GetRecordRequestBody}
\format{
An \code{R6Class} generator object
}
\description{
GetRecordRequestBody Class

GetRecordRequestBody Class
}
\details{
Create a new GetRecordRequestBody
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{select}}{list(character) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetRecordRequestBody-new}{\code{GetRecordRequestBody$new()}}
\item \href{#method-GetRecordRequestBody-toJSON}{\code{GetRecordRequestBody$toJSON()}}
\item \href{#method-GetRecordRequestBody-fromJSON}{\code{GetRecordRequestBody$fromJSON()}}
\item \href{#method-GetRecordRequestBody-toJSONString}{\code{GetRecordRequestBody$toJSONString()}}
\item \href{#method-GetRecordRequestBody-fromJSONString}{\code{GetRecordRequestBody$fromJSONString()}}
\item \href{#method-GetRecordRequestBody-validateJSON}{\code{GetRecordRequestBody$validateJSON()}}
\item \href{#method-GetRecordRequestBody-toString}{\code{GetRecordRequestBody$toString()}}
\item \href{#method-GetRecordRequestBody-isValid}{\code{GetRecordRequestBody$isValid()}}
\item \href{#method-GetRecordRequestBody-getInvalidFields}{\code{GetRecordRequestBody$getInvalidFields()}}
\item \href{#method-GetRecordRequestBody-print}{\code{GetRecordRequestBody$print()}}
\item \href{#method-GetRecordRequestBody-clone}{\code{GetRecordRequestBody$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRecordRequestBody-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetRecordRequestBody-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new GetRecordRequestBody class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetRecordRequestBody$new(select = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{select}}{select}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRecordRequestBody-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GetRecordRequestBody-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetRecordRequestBody$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
GetRecordRequestBody in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRecordRequestBody-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GetRecordRequestBody-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of GetRecordRequestBody
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetRecordRequestBody$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of GetRecordRequestBody
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRecordRequestBody-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GetRecordRequestBody-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetRecordRequestBody$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
GetRecordRequestBody in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRecordRequestBody-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GetRecordRequestBody-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of GetRecordRequestBody
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetRecordRequestBody$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of GetRecordRequestBody
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRecordRequestBody-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GetRecordRequestBody-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to GetRecordRequestBody and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetRecordRequestBody$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRecordRequestBody-toString"></a>}}
\if{latex}{\out{\hypertarget{method-GetRecordRequestBody-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetRecordRequestBody$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of GetRecordRequestBody
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRecordRequestBody-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-GetRecordRequestBody-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetRecordRequestBody$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRecordRequestBody-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-GetRecordRequestBody-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetRecordRequestBody$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRecordRequestBody-print"></a>}}
\if{latex}{\out{\hypertarget{method-GetRecordRequestBody-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetRecordRequestBody$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRecordRequestBody-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetRecordRequestBody-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetRecordRequestBody$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
