% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_records_request_body.R
\docType{class}
\name{GetRecordsRequestBody}
\alias{GetRecordsRequestBody}
\title{GetRecordsRequestBody}
\format{
An \code{R6Class} generator object
}
\description{
GetRecordsRequestBody Class

GetRecordsRequestBody Class
}
\details{
Create a new GetRecordsRequestBody
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{select}}{list(character) [optional]}

\item{\code{filter}}{object [optional]}

\item{\code{order_by}}{list(\link{OrderByColumn}) [optional]}

\item{\code{search}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetRecordsRequestBody-new}{\code{GetRecordsRequestBody$new()}}
\item \href{#method-GetRecordsRequestBody-toJSON}{\code{GetRecordsRequestBody$toJSON()}}
\item \href{#method-GetRecordsRequestBody-fromJSON}{\code{GetRecordsRequestBody$fromJSON()}}
\item \href{#method-GetRecordsRequestBody-toJSONString}{\code{GetRecordsRequestBody$toJSONString()}}
\item \href{#method-GetRecordsRequestBody-fromJSONString}{\code{GetRecordsRequestBody$fromJSONString()}}
\item \href{#method-GetRecordsRequestBody-validateJSON}{\code{GetRecordsRequestBody$validateJSON()}}
\item \href{#method-GetRecordsRequestBody-toString}{\code{GetRecordsRequestBody$toString()}}
\item \href{#method-GetRecordsRequestBody-isValid}{\code{GetRecordsRequestBody$isValid()}}
\item \href{#method-GetRecordsRequestBody-getInvalidFields}{\code{GetRecordsRequestBody$getInvalidFields()}}
\item \href{#method-GetRecordsRequestBody-print}{\code{GetRecordsRequestBody$print()}}
\item \href{#method-GetRecordsRequestBody-clone}{\code{GetRecordsRequestBody$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRecordsRequestBody-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetRecordsRequestBody-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new GetRecordsRequestBody class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetRecordsRequestBody$new(
  select = NULL,
  filter = NULL,
  order_by = NULL,
  search = "",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{select}}{select}

\item{\code{filter}}{filter}

\item{\code{order_by}}{order_by}

\item{\code{search}}{search. Default to "".}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRecordsRequestBody-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GetRecordsRequestBody-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetRecordsRequestBody$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
GetRecordsRequestBody in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRecordsRequestBody-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GetRecordsRequestBody-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of GetRecordsRequestBody
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetRecordsRequestBody$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of GetRecordsRequestBody
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRecordsRequestBody-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GetRecordsRequestBody-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetRecordsRequestBody$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
GetRecordsRequestBody in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRecordsRequestBody-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GetRecordsRequestBody-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of GetRecordsRequestBody
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetRecordsRequestBody$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of GetRecordsRequestBody
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRecordsRequestBody-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GetRecordsRequestBody-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to GetRecordsRequestBody and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetRecordsRequestBody$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRecordsRequestBody-toString"></a>}}
\if{latex}{\out{\hypertarget{method-GetRecordsRequestBody-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetRecordsRequestBody$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of GetRecordsRequestBody
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRecordsRequestBody-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-GetRecordsRequestBody-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetRecordsRequestBody$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRecordsRequestBody-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-GetRecordsRequestBody-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetRecordsRequestBody$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRecordsRequestBody-print"></a>}}
\if{latex}{\out{\hypertarget{method-GetRecordsRequestBody-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetRecordsRequestBody$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRecordsRequestBody-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetRecordsRequestBody-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetRecordsRequestBody$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
