% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_request_body.R
\docType{class}
\name{GroupByRequestBody}
\alias{GroupByRequestBody}
\title{GroupByRequestBody}
\format{
An \code{R6Class} generator object
}
\description{
GroupByRequestBody Class

GroupByRequestBody Class
}
\details{
Create a new GroupByRequestBody
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dimensions}}{list(\link{Dimension})}

\item{\code{measures}}{list(\link{Measure})}

\item{\code{filter}}{object [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GroupByRequestBody-new}{\code{GroupByRequestBody$new()}}
\item \href{#method-GroupByRequestBody-toJSON}{\code{GroupByRequestBody$toJSON()}}
\item \href{#method-GroupByRequestBody-fromJSON}{\code{GroupByRequestBody$fromJSON()}}
\item \href{#method-GroupByRequestBody-toJSONString}{\code{GroupByRequestBody$toJSONString()}}
\item \href{#method-GroupByRequestBody-fromJSONString}{\code{GroupByRequestBody$fromJSONString()}}
\item \href{#method-GroupByRequestBody-validateJSON}{\code{GroupByRequestBody$validateJSON()}}
\item \href{#method-GroupByRequestBody-toString}{\code{GroupByRequestBody$toString()}}
\item \href{#method-GroupByRequestBody-isValid}{\code{GroupByRequestBody$isValid()}}
\item \href{#method-GroupByRequestBody-getInvalidFields}{\code{GroupByRequestBody$getInvalidFields()}}
\item \href{#method-GroupByRequestBody-print}{\code{GroupByRequestBody$print()}}
\item \href{#method-GroupByRequestBody-clone}{\code{GroupByRequestBody$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupByRequestBody-new"></a>}}
\if{latex}{\out{\hypertarget{method-GroupByRequestBody-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new GroupByRequestBody class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupByRequestBody$new(dimensions, measures, filter = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dimensions}}{dimensions}

\item{\code{measures}}{measures}

\item{\code{filter}}{filter}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupByRequestBody-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupByRequestBody-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupByRequestBody$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
GroupByRequestBody in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupByRequestBody-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupByRequestBody-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of GroupByRequestBody
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupByRequestBody$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of GroupByRequestBody
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupByRequestBody-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupByRequestBody-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupByRequestBody$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
GroupByRequestBody in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupByRequestBody-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupByRequestBody-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of GroupByRequestBody
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupByRequestBody$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of GroupByRequestBody
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupByRequestBody-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupByRequestBody-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to GroupByRequestBody and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupByRequestBody$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupByRequestBody-toString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupByRequestBody-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupByRequestBody$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of GroupByRequestBody
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupByRequestBody-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-GroupByRequestBody-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupByRequestBody$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupByRequestBody-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-GroupByRequestBody-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupByRequestBody$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupByRequestBody-print"></a>}}
\if{latex}{\out{\hypertarget{method-GroupByRequestBody-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupByRequestBody$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupByRequestBody-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GroupByRequestBody-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupByRequestBody$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
