% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laminr.api-package.R
\docType{package}
\name{laminr.api-package}
\alias{laminr.api}
\alias{laminr.api-package}
\title{laminr.api: Interface to the 'LaminDB' API}
\description{
Client package for accessing the Lamin API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/data-intuitive/laminr.api}
  \item Report bugs at \url{https://github.com/data-intuitive/laminr.api/issues}
}

}
\author{
\strong{Maintainer}: Robrecht Cannoodt \email{robrecht@data-intuitive.com} (\href{https://orcid.org/0000-0003-3641-729X}{ORCID}) (rcannood)

Authors:
\itemize{
  \item Luke Zappia \email{luke@data-intuitive.com} (\href{https://orcid.org/0000-0001-7744-8565}{ORCID}) (lazappi)
  \item Data Intuitive \email{info@data-intuitive.com}
  \item Lamin Labs \email{open-source@lamin.ai} [copyright holder]
}

}
\keyword{internal}
