% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\docType{class}
\name{Measure}
\alias{Measure}
\title{Measure}
\format{
An \code{R6Class} generator object
}
\description{
Measure Class

Measure Class
}
\details{
Create a new Measure
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{field_name}}{character}

\item{\code{agg_func}}{character}

\item{\code{alias}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Measure-new}{\code{Measure$new()}}
\item \href{#method-Measure-toJSON}{\code{Measure$toJSON()}}
\item \href{#method-Measure-fromJSON}{\code{Measure$fromJSON()}}
\item \href{#method-Measure-toJSONString}{\code{Measure$toJSONString()}}
\item \href{#method-Measure-fromJSONString}{\code{Measure$fromJSONString()}}
\item \href{#method-Measure-validateJSON}{\code{Measure$validateJSON()}}
\item \href{#method-Measure-toString}{\code{Measure$toString()}}
\item \href{#method-Measure-isValid}{\code{Measure$isValid()}}
\item \href{#method-Measure-getInvalidFields}{\code{Measure$getInvalidFields()}}
\item \href{#method-Measure-print}{\code{Measure$print()}}
\item \href{#method-Measure-clone}{\code{Measure$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Measure-new"></a>}}
\if{latex}{\out{\hypertarget{method-Measure-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Measure class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Measure$new(field_name, agg_func, alias = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field_name}}{field_name}

\item{\code{agg_func}}{agg_func}

\item{\code{alias}}{alias}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Measure-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Measure-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Measure$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Measure in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Measure-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Measure-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Measure
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Measure$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Measure
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Measure-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Measure-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Measure$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Measure in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Measure-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Measure-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Measure
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Measure$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Measure
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Measure-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Measure-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Measure and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Measure$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Measure-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Measure-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Measure$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Measure
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Measure-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Measure-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Measure$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Measure-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Measure-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Measure$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Measure-print"></a>}}
\if{latex}{\out{\hypertarget{method-Measure-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Measure$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Measure-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Measure-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Measure$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
