% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_by_column.R
\docType{class}
\name{OrderByColumn}
\alias{OrderByColumn}
\title{OrderByColumn}
\format{
An \code{R6Class} generator object
}
\description{
OrderByColumn Class

OrderByColumn Class
}
\details{
Create a new OrderByColumn
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{field}}{character}

\item{\code{descending}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrderByColumn-new}{\code{OrderByColumn$new()}}
\item \href{#method-OrderByColumn-toJSON}{\code{OrderByColumn$toJSON()}}
\item \href{#method-OrderByColumn-fromJSON}{\code{OrderByColumn$fromJSON()}}
\item \href{#method-OrderByColumn-toJSONString}{\code{OrderByColumn$toJSONString()}}
\item \href{#method-OrderByColumn-fromJSONString}{\code{OrderByColumn$fromJSONString()}}
\item \href{#method-OrderByColumn-validateJSON}{\code{OrderByColumn$validateJSON()}}
\item \href{#method-OrderByColumn-toString}{\code{OrderByColumn$toString()}}
\item \href{#method-OrderByColumn-isValid}{\code{OrderByColumn$isValid()}}
\item \href{#method-OrderByColumn-getInvalidFields}{\code{OrderByColumn$getInvalidFields()}}
\item \href{#method-OrderByColumn-print}{\code{OrderByColumn$print()}}
\item \href{#method-OrderByColumn-clone}{\code{OrderByColumn$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderByColumn-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrderByColumn-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OrderByColumn class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderByColumn$new(field, descending = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{field}

\item{\code{descending}}{descending. Default to FALSE.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderByColumn-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrderByColumn-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderByColumn$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrderByColumn in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderByColumn-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrderByColumn-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OrderByColumn
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderByColumn$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrderByColumn
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderByColumn-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrderByColumn-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderByColumn$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OrderByColumn in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderByColumn-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OrderByColumn-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OrderByColumn
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderByColumn$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OrderByColumn
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderByColumn-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrderByColumn-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OrderByColumn and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderByColumn$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderByColumn-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OrderByColumn-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderByColumn$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OrderByColumn
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderByColumn-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OrderByColumn-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderByColumn$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderByColumn-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OrderByColumn-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderByColumn$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderByColumn-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrderByColumn-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderByColumn$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderByColumn-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrderByColumn-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderByColumn$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
