% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_permissions_request.R
\docType{class}
\name{S3PermissionsRequest}
\alias{S3PermissionsRequest}
\title{S3PermissionsRequest}
\format{
An \code{R6Class} generator object
}
\description{
S3PermissionsRequest Class

S3PermissionsRequest Class
}
\details{
Create a new S3PermissionsRequest
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{aws_access_key_id}}{character}

\item{\code{aws_secret_access_key}}{character}

\item{\code{region}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-S3PermissionsRequest-new}{\code{S3PermissionsRequest$new()}}
\item \href{#method-S3PermissionsRequest-toJSON}{\code{S3PermissionsRequest$toJSON()}}
\item \href{#method-S3PermissionsRequest-fromJSON}{\code{S3PermissionsRequest$fromJSON()}}
\item \href{#method-S3PermissionsRequest-toJSONString}{\code{S3PermissionsRequest$toJSONString()}}
\item \href{#method-S3PermissionsRequest-fromJSONString}{\code{S3PermissionsRequest$fromJSONString()}}
\item \href{#method-S3PermissionsRequest-validateJSON}{\code{S3PermissionsRequest$validateJSON()}}
\item \href{#method-S3PermissionsRequest-toString}{\code{S3PermissionsRequest$toString()}}
\item \href{#method-S3PermissionsRequest-isValid}{\code{S3PermissionsRequest$isValid()}}
\item \href{#method-S3PermissionsRequest-getInvalidFields}{\code{S3PermissionsRequest$getInvalidFields()}}
\item \href{#method-S3PermissionsRequest-print}{\code{S3PermissionsRequest$print()}}
\item \href{#method-S3PermissionsRequest-clone}{\code{S3PermissionsRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3PermissionsRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-S3PermissionsRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new S3PermissionsRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3PermissionsRequest$new(aws_access_key_id, aws_secret_access_key, region, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aws_access_key_id}}{aws_access_key_id}

\item{\code{aws_secret_access_key}}{aws_secret_access_key}

\item{\code{region}}{region}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3PermissionsRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-S3PermissionsRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3PermissionsRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
S3PermissionsRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3PermissionsRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-S3PermissionsRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of S3PermissionsRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3PermissionsRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of S3PermissionsRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3PermissionsRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-S3PermissionsRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3PermissionsRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
S3PermissionsRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3PermissionsRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-S3PermissionsRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of S3PermissionsRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3PermissionsRequest$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of S3PermissionsRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3PermissionsRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-S3PermissionsRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to S3PermissionsRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3PermissionsRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3PermissionsRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-S3PermissionsRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3PermissionsRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of S3PermissionsRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3PermissionsRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-S3PermissionsRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3PermissionsRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3PermissionsRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-S3PermissionsRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3PermissionsRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3PermissionsRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-S3PermissionsRequest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3PermissionsRequest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3PermissionsRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-S3PermissionsRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3PermissionsRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
