% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_error_loc_inner.R
\docType{class}
\name{ValidationErrorLocInner}
\alias{ValidationErrorLocInner}
\title{ValidationErrorLocInner}
\format{
An \code{R6Class} generator object
}
\description{
ValidationErrorLocInner Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{any_of}}{a list of object types defined in the anyOf schema.
Initialize a new ValidationErrorLocInner.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ValidationErrorLocInner-new}{\code{ValidationErrorLocInner$new()}}
\item \href{#method-ValidationErrorLocInner-fromJSONString}{\code{ValidationErrorLocInner$fromJSONString()}}
\item \href{#method-ValidationErrorLocInner-fromJSON}{\code{ValidationErrorLocInner$fromJSON()}}
\item \href{#method-ValidationErrorLocInner-toJSONString}{\code{ValidationErrorLocInner$toJSONString()}}
\item \href{#method-ValidationErrorLocInner-toJSON}{\code{ValidationErrorLocInner$toJSON()}}
\item \href{#method-ValidationErrorLocInner-validateJSON}{\code{ValidationErrorLocInner$validateJSON()}}
\item \href{#method-ValidationErrorLocInner-toString}{\code{ValidationErrorLocInner$toString()}}
\item \href{#method-ValidationErrorLocInner-print}{\code{ValidationErrorLocInner$print()}}
\item \href{#method-ValidationErrorLocInner-clone}{\code{ValidationErrorLocInner$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidationErrorLocInner-new"></a>}}
\if{latex}{\out{\hypertarget{method-ValidationErrorLocInner-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ValidationErrorLocInner.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationErrorLocInner$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the anyOf schemas: "character", "integer"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidationErrorLocInner-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ValidationErrorLocInner-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ValidationErrorLocInner.
An alias to the method `fromJSON`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationErrorLocInner$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of ValidationErrorLocInner.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidationErrorLocInner-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ValidationErrorLocInner-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ValidationErrorLocInner.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationErrorLocInner$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of ValidationErrorLocInner.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidationErrorLocInner-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ValidationErrorLocInner-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize ValidationErrorLocInner to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationErrorLocInner$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the ValidationErrorLocInner.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidationErrorLocInner-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ValidationErrorLocInner-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize ValidationErrorLocInner to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationErrorLocInner$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the ValidationErrorLocInner.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidationErrorLocInner-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ValidationErrorLocInner-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to ValidationErrorLocInner and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationErrorLocInner$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidationErrorLocInner-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ValidationErrorLocInner-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationErrorLocInner$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidationErrorLocInner-print"></a>}}
\if{latex}{\out{\hypertarget{method-ValidationErrorLocInner-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationErrorLocInner$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidationErrorLocInner-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ValidationErrorLocInner-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationErrorLocInner$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
