% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_api.R
\docType{class}
\name{DefaultApi}
\alias{DefaultApi}
\title{Default operations}
\format{
An \code{R6Class} generator object
}
\description{
DefaultApi
}
\details{
FastAPI

No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

The version of the OpenAPI document: 0.1.0
Generated by: https://openapi-generator.tech
}
\examples{
\dontrun{
####################  AddCollaboratorInstancesInstanceIdCollaboratorsAccountIdPut  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_account_id <- "account_id_example" # character |
var_schema_id <- "schema_id_example" # character |
var_role <- "read" # character |  (Optional)
var_authorization <- "authorization_example" # character |  (Optional)

# Add Collaborator
api_instance <- DefaultApi$new()



####################  CreateInstanceInstancesPut  ####################

library(laminr.api)
var_name <- "name_example" # character |
var_storage <- "create-s3" # character |  (Optional)
var_schema_str <- "schema_str_example" # character |  (Optional)
var_db_server_key <- "db_server_key_example" # character |  (Optional)
var_db_server_url <- "db_server_url_example" # character |  (Optional)
var_authorization <- "authorization_example" # character |  (Optional)

# Create Instance
api_instance <- DefaultApi$new()



####################  CreateRecordInstancesInstanceIdModulesModuleNameModelNamePut  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_module_name <- "module_name_example" # character |
var_model_name <- "model_name_example" # character |
var_schema_id <- "schema_id_example" # character |
var_body <- TODO # AnyType |
var_authorization <- "authorization_example" # character |  (Optional)

# Create Record
api_instance <- DefaultApi$new()



####################  DeleteCollaboratorInstancesInstanceIdCollaboratorsAccountIdDelete  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_account_id <- "account_id_example" # character |
var_authorization <- "authorization_example" # character |  (Optional)

# Delete Collaborator
api_instance <- DefaultApi$new()



####################  DeleteInstanceInstancesInstanceIdDelete  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_instance_name <- "instance_name_example" # character |
var_authorization <- "authorization_example" # character |  (Optional)

# Delete Instance
api_instance <- DefaultApi$new()



####################  DeleteRecordInstancesInstanceIdModulesModuleNameModelNameUidDelete  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_module_name <- "module_name_example" # character |
var_model_name <- "model_name_example" # character |
var_uid <- "uid_example" # character |
var_schema_id <- "schema_id_example" # character |
var_authorization <- "authorization_example" # character |  (Optional)

# Delete Record
api_instance <- DefaultApi$new()



####################  GeneratePasswordInstancesInstanceIdDbPasswordGet  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_authorization <- "authorization_example" # character |  (Optional)

# Generate Password
api_instance <- DefaultApi$new()



####################  GenerateUrlInstancesInstanceIdDbUrlGet  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_authorization <- "authorization_example" # character |  (Optional)

# Generate Url
api_instance <- DefaultApi$new()



####################  GetInstanceStatisticsInstancesInstanceIdStatisticsGet  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_schema_id <- "schema_id_example" # character |
var_q <- c("inner_example") # array[character] | In module + "." + model format (case-sensitive) (Optional)
var_authorization <- "authorization_example" # character |  (Optional)

# Get Instance Statistics
api_instance <- DefaultApi$new()



####################  GetNonEmptyTablesInstancesInstanceIdNonEmptyTablesGet  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_schema_id <- "schema_id_example" # character |
var_authorization <- "authorization_example" # character |  (Optional)

# Get Non Empty Tables
api_instance <- DefaultApi$new()



####################  GetRecordInstancesInstanceIdModulesModuleNameModelNameIdOrUidPost  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_module_name <- "module_name_example" # character |
var_model_name <- "model_name_example" # character |
var_id_or_uid <- "id_or_uid_example" # character |
var_schema_id <- "schema_id_example" # character |
var_limit_to_many <- 10 # integer |  (Optional)
var_include_foreign_keys <- FALSE # character |  (Optional)
var_authorization <- "authorization_example" # character |  (Optional)
var_get_record_request_body <- GetRecordRequestBody$new(c("select_example")) # GetRecordRequestBody |  (Optional)

# Get Record
api_instance <- DefaultApi$new()



####################  GetRecordsInstancesInstanceIdModulesModuleNameModelNamePost  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_module_name <- "module_name_example" # character |
var_model_name <- "model_name_example" # character |
var_schema_id <- "schema_id_example" # character |
var_limit <- 50 # integer |  (Optional)
var_offset <- 0 # integer |  (Optional)
var_limit_to_many <- 10 # integer |  (Optional)
var_include_foreign_keys <- FALSE # character |  (Optional)
var_authorization <- "authorization_example" # character |  (Optional)
var_get_records_request_body <- GetRecordsRequestBody$new(c("select_example"), 123, c(OrderByColumn$new("field_example", "descending_example")), "search_example") # GetRecordsRequestBody |  (Optional)

# Get Records
api_instance <- DefaultApi$new()



####################  GetRelationCountsInstancesInstanceIdModulesModuleNameModelNameIdCountsGet  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_module_name <- "module_name_example" # character |
var_model_name <- "model_name_example" # character |
var_id <- 56 # integer |
var_schema_id <- "schema_id_example" # character |
var_authorization <- "authorization_example" # character |  (Optional)

# Get Relation Counts
api_instance <- DefaultApi$new()



####################  GetRelationsInstancesInstanceIdSchemaModuleNameModelNameGet  ####################

library(laminr.api)
var_module_name <- "module_name_example" # character |
var_model_name <- "model_name_example" # character |
var_instance_id <- "instance_id_example" # character |
var_authorization <- "authorization_example" # character |  (Optional)

# Get Relations
api_instance <- DefaultApi$new()



####################  GetSchemaInstancesInstanceIdSchemaGet  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_authorization <- "authorization_example" # character |  (Optional)

# Get Schema
api_instance <- DefaultApi$new()



####################  GetValuesInstancesInstanceIdModulesModuleNameModelNameFieldsFieldPathPost  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_module_name <- "module_name_example" # character |
var_model_name <- "model_name_example" # character |
var_field_path <- "field_path_example" # character |
var_schema_id <- "schema_id_example" # character |
var_limit <- 50 # integer |  (Optional)
var_offset <- 0 # integer |  (Optional)
var_authorization <- "authorization_example" # character |  (Optional)
var_get_values_request_body <- GetValuesRequestBody$new(123, c(OrderByColumn$new("field_example", "descending_example")), "search_example") # GetValuesRequestBody |  (Optional)

# Get Values
api_instance <- DefaultApi$new()



####################  GrantS3PermissionsStoragesS3BucketNamePermissionsPut  ####################

library(laminr.api)
var_bucket_name <- "bucket_name_example" # character |
var_s3_permissions_request <- S3PermissionsRequest$new("aws_access_key_id_example", "aws_secret_access_key_example", "region_example") # S3PermissionsRequest |
var_aws_account_id <- "767398070972" # character |  (Optional)
var_aws_user_name <- "lamin-manager" # character |  (Optional)
var_authorization <- "authorization_example" # character |  (Optional)

# Grant S3 Permissions
api_instance <- DefaultApi$new()



####################  GroupByInstancesInstanceIdModulesModuleNameModelNameGroupByPost  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_module_name <- "module_name_example" # character |
var_model_name <- "model_name_example" # character |
var_schema_id <- "schema_id_example" # character |
var_group_by_request_body <- GroupByRequestBody$new(c(Dimension$new("field_name_example", "count")), c(Measure$new("field_name_example", "count", "alias_example")), 123) # GroupByRequestBody |
var_authorization <- "authorization_example" # character |  (Optional)

# Group By
api_instance <- DefaultApi$new()



####################  TransferOwnershipInstancesInstanceIdOwnerHandlePatch  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_handle <- "handle_example" # character |
var_schema_id <- "schema_id_example" # character |
var_authorization <- "authorization_example" # character |  (Optional)

# Transfer Ownership
api_instance <- DefaultApi$new()



####################  UpdateCollaboratorInstancesInstanceIdCollaboratorsAccountIdPatch  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_account_id <- "account_id_example" # character |
var_role <- "role_example" # character |
var_authorization <- "authorization_example" # character |  (Optional)

# Update Collaborator
api_instance <- DefaultApi$new()



####################  UpdateRecordInstancesInstanceIdModulesModuleNameModelNameUidPatch  ####################

library(laminr.api)
var_instance_id <- "instance_id_example" # character |
var_module_name <- "module_name_example" # character |
var_model_name <- "model_name_example" # character |
var_uid <- "uid_example" # character |
var_schema_id <- "schema_id_example" # character |
var_body <- TODO # AnyType |
var_authorization <- "authorization_example" # character |  (Optional)

# Update Record
api_instance <- DefaultApi$new()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DefaultApi-new}{\code{DefaultApi$new()}}
\item \href{#method-DefaultApi-AddCollaboratorInstancesInstanceIdCollaboratorsAccountIdPut}{\code{DefaultApi$AddCollaboratorInstancesInstanceIdCollaboratorsAccountIdPut()}}
\item \href{#method-DefaultApi-AddCollaboratorInstancesInstanceIdCollaboratorsAccountIdPutWithHttpInfo}{\code{DefaultApi$AddCollaboratorInstancesInstanceIdCollaboratorsAccountIdPutWithHttpInfo()}}
\item \href{#method-DefaultApi-CreateInstanceInstancesPut}{\code{DefaultApi$CreateInstanceInstancesPut()}}
\item \href{#method-DefaultApi-CreateInstanceInstancesPutWithHttpInfo}{\code{DefaultApi$CreateInstanceInstancesPutWithHttpInfo()}}
\item \href{#method-DefaultApi-CreateRecordInstancesInstanceIdModulesModuleNameModelNamePut}{\code{DefaultApi$CreateRecordInstancesInstanceIdModulesModuleNameModelNamePut()}}
\item \href{#method-DefaultApi-CreateRecordInstancesInstanceIdModulesModuleNameModelNamePutWithHttpInfo}{\code{DefaultApi$CreateRecordInstancesInstanceIdModulesModuleNameModelNamePutWithHttpInfo()}}
\item \href{#method-DefaultApi-DeleteCollaboratorInstancesInstanceIdCollaboratorsAccountIdDelete}{\code{DefaultApi$DeleteCollaboratorInstancesInstanceIdCollaboratorsAccountIdDelete()}}
\item \href{#method-DefaultApi-DeleteCollaboratorInstancesInstanceIdCollaboratorsAccountIdDeleteWithHttpInfo}{\code{DefaultApi$DeleteCollaboratorInstancesInstanceIdCollaboratorsAccountIdDeleteWithHttpInfo()}}
\item \href{#method-DefaultApi-DeleteInstanceInstancesInstanceIdDelete}{\code{DefaultApi$DeleteInstanceInstancesInstanceIdDelete()}}
\item \href{#method-DefaultApi-DeleteInstanceInstancesInstanceIdDeleteWithHttpInfo}{\code{DefaultApi$DeleteInstanceInstancesInstanceIdDeleteWithHttpInfo()}}
\item \href{#method-DefaultApi-DeleteRecordInstancesInstanceIdModulesModuleNameModelNameUidDelete}{\code{DefaultApi$DeleteRecordInstancesInstanceIdModulesModuleNameModelNameUidDelete()}}
\item \href{#method-DefaultApi-DeleteRecordInstancesInstanceIdModulesModuleNameModelNameUidDeleteWithHttpInfo}{\code{DefaultApi$DeleteRecordInstancesInstanceIdModulesModuleNameModelNameUidDeleteWithHttpInfo()}}
\item \href{#method-DefaultApi-GeneratePasswordInstancesInstanceIdDbPasswordGet}{\code{DefaultApi$GeneratePasswordInstancesInstanceIdDbPasswordGet()}}
\item \href{#method-DefaultApi-GeneratePasswordInstancesInstanceIdDbPasswordGetWithHttpInfo}{\code{DefaultApi$GeneratePasswordInstancesInstanceIdDbPasswordGetWithHttpInfo()}}
\item \href{#method-DefaultApi-GenerateUrlInstancesInstanceIdDbUrlGet}{\code{DefaultApi$GenerateUrlInstancesInstanceIdDbUrlGet()}}
\item \href{#method-DefaultApi-GenerateUrlInstancesInstanceIdDbUrlGetWithHttpInfo}{\code{DefaultApi$GenerateUrlInstancesInstanceIdDbUrlGetWithHttpInfo()}}
\item \href{#method-DefaultApi-GetInstanceStatisticsInstancesInstanceIdStatisticsGet}{\code{DefaultApi$GetInstanceStatisticsInstancesInstanceIdStatisticsGet()}}
\item \href{#method-DefaultApi-GetInstanceStatisticsInstancesInstanceIdStatisticsGetWithHttpInfo}{\code{DefaultApi$GetInstanceStatisticsInstancesInstanceIdStatisticsGetWithHttpInfo()}}
\item \href{#method-DefaultApi-GetNonEmptyTablesInstancesInstanceIdNonEmptyTablesGet}{\code{DefaultApi$GetNonEmptyTablesInstancesInstanceIdNonEmptyTablesGet()}}
\item \href{#method-DefaultApi-GetNonEmptyTablesInstancesInstanceIdNonEmptyTablesGetWithHttpInfo}{\code{DefaultApi$GetNonEmptyTablesInstancesInstanceIdNonEmptyTablesGetWithHttpInfo()}}
\item \href{#method-DefaultApi-GetRecordInstancesInstanceIdModulesModuleNameModelNameIdOrUidPost}{\code{DefaultApi$GetRecordInstancesInstanceIdModulesModuleNameModelNameIdOrUidPost()}}
\item \href{#method-DefaultApi-GetRecordInstancesInstanceIdModulesModuleNameModelNameIdOrUidPostWithHttpInfo}{\code{DefaultApi$GetRecordInstancesInstanceIdModulesModuleNameModelNameIdOrUidPostWithHttpInfo()}}
\item \href{#method-DefaultApi-GetRecordsInstancesInstanceIdModulesModuleNameModelNamePost}{\code{DefaultApi$GetRecordsInstancesInstanceIdModulesModuleNameModelNamePost()}}
\item \href{#method-DefaultApi-GetRecordsInstancesInstanceIdModulesModuleNameModelNamePostWithHttpInfo}{\code{DefaultApi$GetRecordsInstancesInstanceIdModulesModuleNameModelNamePostWithHttpInfo()}}
\item \href{#method-DefaultApi-GetRelationCountsInstancesInstanceIdModulesModuleNameModelNameIdCountsGet}{\code{DefaultApi$GetRelationCountsInstancesInstanceIdModulesModuleNameModelNameIdCountsGet()}}
\item \href{#method-DefaultApi-GetRelationCountsInstancesInstanceIdModulesModuleNameModelNameIdCountsGetWithHttpInfo}{\code{DefaultApi$GetRelationCountsInstancesInstanceIdModulesModuleNameModelNameIdCountsGetWithHttpInfo()}}
\item \href{#method-DefaultApi-GetRelationsInstancesInstanceIdSchemaModuleNameModelNameGet}{\code{DefaultApi$GetRelationsInstancesInstanceIdSchemaModuleNameModelNameGet()}}
\item \href{#method-DefaultApi-GetRelationsInstancesInstanceIdSchemaModuleNameModelNameGetWithHttpInfo}{\code{DefaultApi$GetRelationsInstancesInstanceIdSchemaModuleNameModelNameGetWithHttpInfo()}}
\item \href{#method-DefaultApi-GetSchemaInstancesInstanceIdSchemaGet}{\code{DefaultApi$GetSchemaInstancesInstanceIdSchemaGet()}}
\item \href{#method-DefaultApi-GetSchemaInstancesInstanceIdSchemaGetWithHttpInfo}{\code{DefaultApi$GetSchemaInstancesInstanceIdSchemaGetWithHttpInfo()}}
\item \href{#method-DefaultApi-GetValuesInstancesInstanceIdModulesModuleNameModelNameFieldsFieldPathPost}{\code{DefaultApi$GetValuesInstancesInstanceIdModulesModuleNameModelNameFieldsFieldPathPost()}}
\item \href{#method-DefaultApi-GetValuesInstancesInstanceIdModulesModuleNameModelNameFieldsFieldPathPostWithHttpInfo}{\code{DefaultApi$GetValuesInstancesInstanceIdModulesModuleNameModelNameFieldsFieldPathPostWithHttpInfo()}}
\item \href{#method-DefaultApi-GrantS3PermissionsStoragesS3BucketNamePermissionsPut}{\code{DefaultApi$GrantS3PermissionsStoragesS3BucketNamePermissionsPut()}}
\item \href{#method-DefaultApi-GrantS3PermissionsStoragesS3BucketNamePermissionsPutWithHttpInfo}{\code{DefaultApi$GrantS3PermissionsStoragesS3BucketNamePermissionsPutWithHttpInfo()}}
\item \href{#method-DefaultApi-GroupByInstancesInstanceIdModulesModuleNameModelNameGroupByPost}{\code{DefaultApi$GroupByInstancesInstanceIdModulesModuleNameModelNameGroupByPost()}}
\item \href{#method-DefaultApi-GroupByInstancesInstanceIdModulesModuleNameModelNameGroupByPostWithHttpInfo}{\code{DefaultApi$GroupByInstancesInstanceIdModulesModuleNameModelNameGroupByPostWithHttpInfo()}}
\item \href{#method-DefaultApi-TransferOwnershipInstancesInstanceIdOwnerHandlePatch}{\code{DefaultApi$TransferOwnershipInstancesInstanceIdOwnerHandlePatch()}}
\item \href{#method-DefaultApi-TransferOwnershipInstancesInstanceIdOwnerHandlePatchWithHttpInfo}{\code{DefaultApi$TransferOwnershipInstancesInstanceIdOwnerHandlePatchWithHttpInfo()}}
\item \href{#method-DefaultApi-UpdateCollaboratorInstancesInstanceIdCollaboratorsAccountIdPatch}{\code{DefaultApi$UpdateCollaboratorInstancesInstanceIdCollaboratorsAccountIdPatch()}}
\item \href{#method-DefaultApi-UpdateCollaboratorInstancesInstanceIdCollaboratorsAccountIdPatchWithHttpInfo}{\code{DefaultApi$UpdateCollaboratorInstancesInstanceIdCollaboratorsAccountIdPatchWithHttpInfo()}}
\item \href{#method-DefaultApi-UpdateRecordInstancesInstanceIdModulesModuleNameModelNameUidPatch}{\code{DefaultApi$UpdateRecordInstancesInstanceIdModulesModuleNameModelNameUidPatch()}}
\item \href{#method-DefaultApi-UpdateRecordInstancesInstanceIdModulesModuleNameModelNameUidPatchWithHttpInfo}{\code{DefaultApi$UpdateRecordInstancesInstanceIdModulesModuleNameModelNameUidPatchWithHttpInfo()}}
\item \href{#method-DefaultApi-clone}{\code{DefaultApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DefaultApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-AddCollaboratorInstancesInstanceIdCollaboratorsAccountIdPut"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-AddCollaboratorInstancesInstanceIdCollaboratorsAccountIdPut}{}}}
\subsection{Method \code{AddCollaboratorInstancesInstanceIdCollaboratorsAccountIdPut()}}{
Add Collaborator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$AddCollaboratorInstancesInstanceIdCollaboratorsAccountIdPut(
  instance_id,
  account_id,
  schema_id,
  role = "read",
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{role}}{(optional) No description (default value: "read")}

\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-AddCollaboratorInstancesInstanceIdCollaboratorsAccountIdPutWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-AddCollaboratorInstancesInstanceIdCollaboratorsAccountIdPutWithHttpInfo}{}}}
\subsection{Method \code{AddCollaboratorInstancesInstanceIdCollaboratorsAccountIdPutWithHttpInfo()}}{
Add Collaborator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$AddCollaboratorInstancesInstanceIdCollaboratorsAccountIdPutWithHttpInfo(
  instance_id,
  account_id,
  schema_id,
  role = "read",
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{role}}{(optional) No description (default value: "read")}

\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-CreateInstanceInstancesPut"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-CreateInstanceInstancesPut}{}}}
\subsection{Method \code{CreateInstanceInstancesPut()}}{
Create Instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$CreateInstanceInstancesPut(
  name,
  storage = "create-s3",
  schema_str = NULL,
  db_server_key = NULL,
  db_server_url = NULL,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{storage}}{(optional) No description (default value: "create-s3")}

\item{\code{schema_str}}{(optional) No description}

\item{\code{db_server_key}}{(optional) No description}

\item{\code{db_server_url}}{(optional) No description}

\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-CreateInstanceInstancesPutWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-CreateInstanceInstancesPutWithHttpInfo}{}}}
\subsection{Method \code{CreateInstanceInstancesPutWithHttpInfo()}}{
Create Instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$CreateInstanceInstancesPutWithHttpInfo(
  name,
  storage = "create-s3",
  schema_str = NULL,
  db_server_key = NULL,
  db_server_url = NULL,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{storage}}{(optional) No description (default value: "create-s3")}

\item{\code{schema_str}}{(optional) No description}

\item{\code{db_server_key}}{(optional) No description}

\item{\code{db_server_url}}{(optional) No description}

\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-CreateRecordInstancesInstanceIdModulesModuleNameModelNamePut"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-CreateRecordInstancesInstanceIdModulesModuleNameModelNamePut}{}}}
\subsection{Method \code{CreateRecordInstancesInstanceIdModulesModuleNameModelNamePut()}}{
Create Record
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$CreateRecordInstancesInstanceIdModulesModuleNameModelNamePut(
  instance_id,
  module_name,
  model_name,
  schema_id,
  body,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-CreateRecordInstancesInstanceIdModulesModuleNameModelNamePutWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-CreateRecordInstancesInstanceIdModulesModuleNameModelNamePutWithHttpInfo}{}}}
\subsection{Method \code{CreateRecordInstancesInstanceIdModulesModuleNameModelNamePutWithHttpInfo()}}{
Create Record
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$CreateRecordInstancesInstanceIdModulesModuleNameModelNamePutWithHttpInfo(
  instance_id,
  module_name,
  model_name,
  schema_id,
  body,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-DeleteCollaboratorInstancesInstanceIdCollaboratorsAccountIdDelete"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-DeleteCollaboratorInstancesInstanceIdCollaboratorsAccountIdDelete}{}}}
\subsection{Method \code{DeleteCollaboratorInstancesInstanceIdCollaboratorsAccountIdDelete()}}{
Delete Collaborator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$DeleteCollaboratorInstancesInstanceIdCollaboratorsAccountIdDelete(
  instance_id,
  account_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-DeleteCollaboratorInstancesInstanceIdCollaboratorsAccountIdDeleteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-DeleteCollaboratorInstancesInstanceIdCollaboratorsAccountIdDeleteWithHttpInfo}{}}}
\subsection{Method \code{DeleteCollaboratorInstancesInstanceIdCollaboratorsAccountIdDeleteWithHttpInfo()}}{
Delete Collaborator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$DeleteCollaboratorInstancesInstanceIdCollaboratorsAccountIdDeleteWithHttpInfo(
  instance_id,
  account_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-DeleteInstanceInstancesInstanceIdDelete"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-DeleteInstanceInstancesInstanceIdDelete}{}}}
\subsection{Method \code{DeleteInstanceInstancesInstanceIdDelete()}}{
Delete Instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$DeleteInstanceInstancesInstanceIdDelete(
  instance_id,
  instance_name,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-DeleteInstanceInstancesInstanceIdDeleteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-DeleteInstanceInstancesInstanceIdDeleteWithHttpInfo}{}}}
\subsection{Method \code{DeleteInstanceInstancesInstanceIdDeleteWithHttpInfo()}}{
Delete Instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$DeleteInstanceInstancesInstanceIdDeleteWithHttpInfo(
  instance_id,
  instance_name,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-DeleteRecordInstancesInstanceIdModulesModuleNameModelNameUidDelete"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-DeleteRecordInstancesInstanceIdModulesModuleNameModelNameUidDelete}{}}}
\subsection{Method \code{DeleteRecordInstancesInstanceIdModulesModuleNameModelNameUidDelete()}}{
Delete Record
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$DeleteRecordInstancesInstanceIdModulesModuleNameModelNameUidDelete(
  instance_id,
  module_name,
  model_name,
  uid,
  schema_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-DeleteRecordInstancesInstanceIdModulesModuleNameModelNameUidDeleteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-DeleteRecordInstancesInstanceIdModulesModuleNameModelNameUidDeleteWithHttpInfo}{}}}
\subsection{Method \code{DeleteRecordInstancesInstanceIdModulesModuleNameModelNameUidDeleteWithHttpInfo()}}{
Delete Record
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$DeleteRecordInstancesInstanceIdModulesModuleNameModelNameUidDeleteWithHttpInfo(
  instance_id,
  module_name,
  model_name,
  uid,
  schema_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GeneratePasswordInstancesInstanceIdDbPasswordGet"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GeneratePasswordInstancesInstanceIdDbPasswordGet}{}}}
\subsection{Method \code{GeneratePasswordInstancesInstanceIdDbPasswordGet()}}{
Generate Password
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GeneratePasswordInstancesInstanceIdDbPasswordGet(
  instance_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GeneratePasswordInstancesInstanceIdDbPasswordGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GeneratePasswordInstancesInstanceIdDbPasswordGetWithHttpInfo}{}}}
\subsection{Method \code{GeneratePasswordInstancesInstanceIdDbPasswordGetWithHttpInfo()}}{
Generate Password
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GeneratePasswordInstancesInstanceIdDbPasswordGetWithHttpInfo(
  instance_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GenerateUrlInstancesInstanceIdDbUrlGet"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GenerateUrlInstancesInstanceIdDbUrlGet}{}}}
\subsection{Method \code{GenerateUrlInstancesInstanceIdDbUrlGet()}}{
Generate Url
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GenerateUrlInstancesInstanceIdDbUrlGet(
  instance_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GenerateUrlInstancesInstanceIdDbUrlGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GenerateUrlInstancesInstanceIdDbUrlGetWithHttpInfo}{}}}
\subsection{Method \code{GenerateUrlInstancesInstanceIdDbUrlGetWithHttpInfo()}}{
Generate Url
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GenerateUrlInstancesInstanceIdDbUrlGetWithHttpInfo(
  instance_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetInstanceStatisticsInstancesInstanceIdStatisticsGet"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetInstanceStatisticsInstancesInstanceIdStatisticsGet}{}}}
\subsection{Method \code{GetInstanceStatisticsInstancesInstanceIdStatisticsGet()}}{
Get Instance Statistics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetInstanceStatisticsInstancesInstanceIdStatisticsGet(
  instance_id,
  schema_id,
  q = NULL,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q}}{(optional) In module + "." + model format (case-sensitive)}

\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetInstanceStatisticsInstancesInstanceIdStatisticsGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetInstanceStatisticsInstancesInstanceIdStatisticsGetWithHttpInfo}{}}}
\subsection{Method \code{GetInstanceStatisticsInstancesInstanceIdStatisticsGetWithHttpInfo()}}{
Get Instance Statistics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetInstanceStatisticsInstancesInstanceIdStatisticsGetWithHttpInfo(
  instance_id,
  schema_id,
  q = NULL,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q}}{(optional) In module + "." + model format (case-sensitive)}

\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetNonEmptyTablesInstancesInstanceIdNonEmptyTablesGet"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetNonEmptyTablesInstancesInstanceIdNonEmptyTablesGet}{}}}
\subsection{Method \code{GetNonEmptyTablesInstancesInstanceIdNonEmptyTablesGet()}}{
Get Non Empty Tables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetNonEmptyTablesInstancesInstanceIdNonEmptyTablesGet(
  instance_id,
  schema_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetNonEmptyTablesInstancesInstanceIdNonEmptyTablesGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetNonEmptyTablesInstancesInstanceIdNonEmptyTablesGetWithHttpInfo}{}}}
\subsection{Method \code{GetNonEmptyTablesInstancesInstanceIdNonEmptyTablesGetWithHttpInfo()}}{
Get Non Empty Tables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetNonEmptyTablesInstancesInstanceIdNonEmptyTablesGetWithHttpInfo(
  instance_id,
  schema_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetRecordInstancesInstanceIdModulesModuleNameModelNameIdOrUidPost"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetRecordInstancesInstanceIdModulesModuleNameModelNameIdOrUidPost}{}}}
\subsection{Method \code{GetRecordInstancesInstanceIdModulesModuleNameModelNameIdOrUidPost()}}{
Get Record
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetRecordInstancesInstanceIdModulesModuleNameModelNameIdOrUidPost(
  instance_id,
  module_name,
  model_name,
  id_or_uid,
  schema_id,
  limit_to_many = 10,
  include_foreign_keys = FALSE,
  authorization = NULL,
  get_record_request_body = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit_to_many}}{(optional) No description (default value: 10)}

\item{\code{include_foreign_keys}}{(optional) No description (default value: FALSE)}

\item{\code{authorization}}{(optional) No description}

\item{\code{get_record_request_body}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetRecordInstancesInstanceIdModulesModuleNameModelNameIdOrUidPostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetRecordInstancesInstanceIdModulesModuleNameModelNameIdOrUidPostWithHttpInfo}{}}}
\subsection{Method \code{GetRecordInstancesInstanceIdModulesModuleNameModelNameIdOrUidPostWithHttpInfo()}}{
Get Record
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetRecordInstancesInstanceIdModulesModuleNameModelNameIdOrUidPostWithHttpInfo(
  instance_id,
  module_name,
  model_name,
  id_or_uid,
  schema_id,
  limit_to_many = 10,
  include_foreign_keys = FALSE,
  authorization = NULL,
  get_record_request_body = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit_to_many}}{(optional) No description (default value: 10)}

\item{\code{include_foreign_keys}}{(optional) No description (default value: FALSE)}

\item{\code{authorization}}{(optional) No description}

\item{\code{get_record_request_body}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetRecordsInstancesInstanceIdModulesModuleNameModelNamePost"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetRecordsInstancesInstanceIdModulesModuleNameModelNamePost}{}}}
\subsection{Method \code{GetRecordsInstancesInstanceIdModulesModuleNameModelNamePost()}}{
Get Records
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetRecordsInstancesInstanceIdModulesModuleNameModelNamePost(
  instance_id,
  module_name,
  model_name,
  schema_id,
  limit = 50,
  offset = 0,
  limit_to_many = 10,
  include_foreign_keys = FALSE,
  authorization = NULL,
  get_records_request_body = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{(optional) No description (default value: 50)}

\item{\code{offset}}{(optional) No description (default value: 0)}

\item{\code{limit_to_many}}{(optional) No description (default value: 10)}

\item{\code{include_foreign_keys}}{(optional) No description (default value: FALSE)}

\item{\code{authorization}}{(optional) No description}

\item{\code{get_records_request_body}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetRecordsInstancesInstanceIdModulesModuleNameModelNamePostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetRecordsInstancesInstanceIdModulesModuleNameModelNamePostWithHttpInfo}{}}}
\subsection{Method \code{GetRecordsInstancesInstanceIdModulesModuleNameModelNamePostWithHttpInfo()}}{
Get Records
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetRecordsInstancesInstanceIdModulesModuleNameModelNamePostWithHttpInfo(
  instance_id,
  module_name,
  model_name,
  schema_id,
  limit = 50,
  offset = 0,
  limit_to_many = 10,
  include_foreign_keys = FALSE,
  authorization = NULL,
  get_records_request_body = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{(optional) No description (default value: 50)}

\item{\code{offset}}{(optional) No description (default value: 0)}

\item{\code{limit_to_many}}{(optional) No description (default value: 10)}

\item{\code{include_foreign_keys}}{(optional) No description (default value: FALSE)}

\item{\code{authorization}}{(optional) No description}

\item{\code{get_records_request_body}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetRelationCountsInstancesInstanceIdModulesModuleNameModelNameIdCountsGet"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetRelationCountsInstancesInstanceIdModulesModuleNameModelNameIdCountsGet}{}}}
\subsection{Method \code{GetRelationCountsInstancesInstanceIdModulesModuleNameModelNameIdCountsGet()}}{
Get Relation Counts
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetRelationCountsInstancesInstanceIdModulesModuleNameModelNameIdCountsGet(
  instance_id,
  module_name,
  model_name,
  id,
  schema_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetRelationCountsInstancesInstanceIdModulesModuleNameModelNameIdCountsGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetRelationCountsInstancesInstanceIdModulesModuleNameModelNameIdCountsGetWithHttpInfo}{}}}
\subsection{Method \code{GetRelationCountsInstancesInstanceIdModulesModuleNameModelNameIdCountsGetWithHttpInfo()}}{
Get Relation Counts
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetRelationCountsInstancesInstanceIdModulesModuleNameModelNameIdCountsGetWithHttpInfo(
  instance_id,
  module_name,
  model_name,
  id,
  schema_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetRelationsInstancesInstanceIdSchemaModuleNameModelNameGet"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetRelationsInstancesInstanceIdSchemaModuleNameModelNameGet}{}}}
\subsection{Method \code{GetRelationsInstancesInstanceIdSchemaModuleNameModelNameGet()}}{
Get Relations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetRelationsInstancesInstanceIdSchemaModuleNameModelNameGet(
  module_name,
  model_name,
  instance_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetRelationsInstancesInstanceIdSchemaModuleNameModelNameGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetRelationsInstancesInstanceIdSchemaModuleNameModelNameGetWithHttpInfo}{}}}
\subsection{Method \code{GetRelationsInstancesInstanceIdSchemaModuleNameModelNameGetWithHttpInfo()}}{
Get Relations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetRelationsInstancesInstanceIdSchemaModuleNameModelNameGetWithHttpInfo(
  module_name,
  model_name,
  instance_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetSchemaInstancesInstanceIdSchemaGet"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetSchemaInstancesInstanceIdSchemaGet}{}}}
\subsection{Method \code{GetSchemaInstancesInstanceIdSchemaGet()}}{
Get Schema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetSchemaInstancesInstanceIdSchemaGet(
  instance_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetSchemaInstancesInstanceIdSchemaGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetSchemaInstancesInstanceIdSchemaGetWithHttpInfo}{}}}
\subsection{Method \code{GetSchemaInstancesInstanceIdSchemaGetWithHttpInfo()}}{
Get Schema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetSchemaInstancesInstanceIdSchemaGetWithHttpInfo(
  instance_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetValuesInstancesInstanceIdModulesModuleNameModelNameFieldsFieldPathPost"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetValuesInstancesInstanceIdModulesModuleNameModelNameFieldsFieldPathPost}{}}}
\subsection{Method \code{GetValuesInstancesInstanceIdModulesModuleNameModelNameFieldsFieldPathPost()}}{
Get Values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetValuesInstancesInstanceIdModulesModuleNameModelNameFieldsFieldPathPost(
  instance_id,
  module_name,
  model_name,
  field_path,
  schema_id,
  limit = 50,
  offset = 0,
  authorization = NULL,
  get_values_request_body = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{(optional) No description (default value: 50)}

\item{\code{offset}}{(optional) No description (default value: 0)}

\item{\code{authorization}}{(optional) No description}

\item{\code{get_values_request_body}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GetValuesInstancesInstanceIdModulesModuleNameModelNameFieldsFieldPathPostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GetValuesInstancesInstanceIdModulesModuleNameModelNameFieldsFieldPathPostWithHttpInfo}{}}}
\subsection{Method \code{GetValuesInstancesInstanceIdModulesModuleNameModelNameFieldsFieldPathPostWithHttpInfo()}}{
Get Values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetValuesInstancesInstanceIdModulesModuleNameModelNameFieldsFieldPathPostWithHttpInfo(
  instance_id,
  module_name,
  model_name,
  field_path,
  schema_id,
  limit = 50,
  offset = 0,
  authorization = NULL,
  get_values_request_body = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{(optional) No description (default value: 50)}

\item{\code{offset}}{(optional) No description (default value: 0)}

\item{\code{authorization}}{(optional) No description}

\item{\code{get_values_request_body}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GrantS3PermissionsStoragesS3BucketNamePermissionsPut"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GrantS3PermissionsStoragesS3BucketNamePermissionsPut}{}}}
\subsection{Method \code{GrantS3PermissionsStoragesS3BucketNamePermissionsPut()}}{
Grant S3 Permissions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GrantS3PermissionsStoragesS3BucketNamePermissionsPut(
  bucket_name,
  s3_permissions_request,
  aws_account_id = "767398070972",
  aws_user_name = "lamin-manager",
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aws_account_id}}{(optional) No description (default value: "767398070972")}

\item{\code{aws_user_name}}{(optional) No description (default value: "lamin-manager")}

\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GrantS3PermissionsStoragesS3BucketNamePermissionsPutWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GrantS3PermissionsStoragesS3BucketNamePermissionsPutWithHttpInfo}{}}}
\subsection{Method \code{GrantS3PermissionsStoragesS3BucketNamePermissionsPutWithHttpInfo()}}{
Grant S3 Permissions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GrantS3PermissionsStoragesS3BucketNamePermissionsPutWithHttpInfo(
  bucket_name,
  s3_permissions_request,
  aws_account_id = "767398070972",
  aws_user_name = "lamin-manager",
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aws_account_id}}{(optional) No description (default value: "767398070972")}

\item{\code{aws_user_name}}{(optional) No description (default value: "lamin-manager")}

\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GroupByInstancesInstanceIdModulesModuleNameModelNameGroupByPost"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GroupByInstancesInstanceIdModulesModuleNameModelNameGroupByPost}{}}}
\subsection{Method \code{GroupByInstancesInstanceIdModulesModuleNameModelNameGroupByPost()}}{
Group By
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GroupByInstancesInstanceIdModulesModuleNameModelNameGroupByPost(
  instance_id,
  module_name,
  model_name,
  schema_id,
  group_by_request_body,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-GroupByInstancesInstanceIdModulesModuleNameModelNameGroupByPostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-GroupByInstancesInstanceIdModulesModuleNameModelNameGroupByPostWithHttpInfo}{}}}
\subsection{Method \code{GroupByInstancesInstanceIdModulesModuleNameModelNameGroupByPostWithHttpInfo()}}{
Group By
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GroupByInstancesInstanceIdModulesModuleNameModelNameGroupByPostWithHttpInfo(
  instance_id,
  module_name,
  model_name,
  schema_id,
  group_by_request_body,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-TransferOwnershipInstancesInstanceIdOwnerHandlePatch"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-TransferOwnershipInstancesInstanceIdOwnerHandlePatch}{}}}
\subsection{Method \code{TransferOwnershipInstancesInstanceIdOwnerHandlePatch()}}{
Transfer Ownership
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$TransferOwnershipInstancesInstanceIdOwnerHandlePatch(
  instance_id,
  handle,
  schema_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-TransferOwnershipInstancesInstanceIdOwnerHandlePatchWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-TransferOwnershipInstancesInstanceIdOwnerHandlePatchWithHttpInfo}{}}}
\subsection{Method \code{TransferOwnershipInstancesInstanceIdOwnerHandlePatchWithHttpInfo()}}{
Transfer Ownership
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$TransferOwnershipInstancesInstanceIdOwnerHandlePatchWithHttpInfo(
  instance_id,
  handle,
  schema_id,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-UpdateCollaboratorInstancesInstanceIdCollaboratorsAccountIdPatch"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-UpdateCollaboratorInstancesInstanceIdCollaboratorsAccountIdPatch}{}}}
\subsection{Method \code{UpdateCollaboratorInstancesInstanceIdCollaboratorsAccountIdPatch()}}{
Update Collaborator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$UpdateCollaboratorInstancesInstanceIdCollaboratorsAccountIdPatch(
  instance_id,
  account_id,
  role,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-UpdateCollaboratorInstancesInstanceIdCollaboratorsAccountIdPatchWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-UpdateCollaboratorInstancesInstanceIdCollaboratorsAccountIdPatchWithHttpInfo}{}}}
\subsection{Method \code{UpdateCollaboratorInstancesInstanceIdCollaboratorsAccountIdPatchWithHttpInfo()}}{
Update Collaborator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$UpdateCollaboratorInstancesInstanceIdCollaboratorsAccountIdPatchWithHttpInfo(
  instance_id,
  account_id,
  role,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-UpdateRecordInstancesInstanceIdModulesModuleNameModelNameUidPatch"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-UpdateRecordInstancesInstanceIdModulesModuleNameModelNameUidPatch}{}}}
\subsection{Method \code{UpdateRecordInstancesInstanceIdModulesModuleNameModelNameUidPatch()}}{
Update Record
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$UpdateRecordInstancesInstanceIdModulesModuleNameModelNameUidPatch(
  instance_id,
  module_name,
  model_name,
  uid,
  schema_id,
  body,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AnyType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-UpdateRecordInstancesInstanceIdModulesModuleNameModelNameUidPatchWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-UpdateRecordInstancesInstanceIdModulesModuleNameModelNameUidPatchWithHttpInfo}{}}}
\subsection{Method \code{UpdateRecordInstancesInstanceIdModulesModuleNameModelNameUidPatchWithHttpInfo()}}{
Update Record
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$UpdateRecordInstancesInstanceIdModulesModuleNameModelNameUidPatchWithHttpInfo(
  instance_id,
  module_name,
  model_name,
  uid,
  schema_id,
  body,
  authorization = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authorization}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (AnyType) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DefaultApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DefaultApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
